krgrpc: Use *_bit() on rpc_desc_recv->flags
Closing connections will need to modify it concurrently to the desc's
thread.

+++ net/krgrpc/rpc_internal.h
@@ -46,7 +46,7 @@ struct rpc_desc_recv {
 	struct list_head list_signal_head;
 	struct rpc_desc_elem *iter;
 	struct rpc_desc_elem *iter_provided;
-	int flags;
+	unsigned long flags;		     /* bitfield */
 };
 
 struct rpc_connection {
